% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_gears.R
\name{add_gear_colors}
\alias{add_gear_colors}
\title{Adds a new gear color configuration}
\usage{
add_gear_colors(gear_code, fill, outline = darken(fill, 0.2))
}
\arguments{
\item{gear_code}{A gear code}

\item{fill}{Its fill color}

\item{outline}{Its outline color}
}
\description{
Adds a new gear color configuration
}
