% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_species.R
\name{add_species_colors}
\alias{add_species_colors}
\title{Adds a new species color configuration}
\usage{
add_species_colors(species_code, fill, outline = darken(fill, 0.2))
}
\arguments{
\item{species_code}{A species code}

\item{fill}{Its fill color}

\item{outline}{Its outline color}
}
\description{
Adds a new species color configuration
}
