% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_fate_types.R
\name{all_fate_type_colors}
\alias{all_fate_type_colors}
\title{Returns the fill / outline colors for all fate types}
\usage{
all_fate_type_colors()
}
\value{
a multiple rows data frame with three columns: \code{fate_type_code}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for all fate types
}
