% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_fishery_types.R
\name{all_fishery_type_colors}
\alias{all_fishery_type_colors}
\title{Returns the fill / outline colors for all fishery types}
\usage{
all_fishery_type_colors()
}
\value{
a multiple rows data frame with three columns: \code{FISHERY_TYPE_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for all fishery types
}
