% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_fishing_grounds.R
\name{all_fishing_ground_colors}
\alias{all_fishing_ground_colors}
\title{Returns the fill / outline colors for all fishing grounds}
\usage{
all_fishing_ground_colors()
}
\value{
a multiple rows data frame with three columns: \code{FISHING_GROUND_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for all fishing grounds
}
