% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_fleets.R
\name{all_fleet_colors}
\alias{all_fleet_colors}
\title{Returns the fill / outline colors for all fleets}
\usage{
all_fleet_colors()
}
\value{
a multiple rows data frame with three columns: \code{FLEET_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for all fleets
}
