% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_gears.R
\name{all_gear_colors}
\alias{all_gear_colors}
\title{Returns the fill / outline colors for the 16 IOTC gear}
\usage{
all_gear_colors()
}
\value{
a multiple rows data frame with three columns: \code{GEAR_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for the 16 IOTC gear
}
