% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_quality_codes.R
\name{all_quality_colors}
\alias{all_quality_colors}
\title{Returns the fill / outline colors for all quality codes}
\usage{
all_quality_colors()
}
\value{
a multiple rows data frame with three columns: \code{QUALITY_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for all quality codes
}
