% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_species_01_group.R
\name{all_species_group_colors}
\alias{all_species_group_colors}
\title{Returns the fill / outline colors for all species groups}
\usage{
all_species_group_colors()
}
\value{
a multiple rows data frame with three columns: \code{SPECIES_GROUP_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for all species groups
}
