% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_helpers.R
\name{color_table}
\alias{color_table}
\title{Builds a data table containing the \code{FILL} and \code{OUTLINE} columns for a
given set of colors, and a \code{FILL_BY} column with color labels.}
\usage{
color_table(colors, labels = NA)
}
\arguments{
\item{colors}{a set of colors (e.g. rainbow(10))}

\item{labels}{the labels for each color. If not provided (i.e., \code{NA}) it is automatically created}
}
\value{
a data table with three columns: \code{FILL_BY}, \code{FILL} and \code{OUTLINE} of which the former is meant to be a label, while the latter two represent fill and outline colors as derived from the provided color inputs
}
\description{
Builds a data table containing the \code{FILL} and \code{OUTLINE} columns for a
given set of colors, and a \code{FILL_BY} column with color labels.
}
