% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_helpers.R
\name{color_table_for}
\alias{color_table_for}
\title{Returns a color table for a given number of unique colors.
Equivalent to \code{color_table(unique_colors(num))}}
\usage{
color_table_for(num_colors)
}
\arguments{
\item{num_colors}{The number of colors to return}
}
\value{
a data table with \code{num_colors} row and three columns: \code{FILL_BY}, \code{FILL} and \code{OUTLINE}. The colors are taken from the first \code{num_colors} entries returned by \code{\link{unique_colors}}
}
\description{
Returns a color table for a given number of unique colors.
Equivalent to \code{color_table(unique_colors(num))}
}
