% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_conditions.R
\name{colors_for_condition}
\alias{colors_for_condition}
\title{Returns the fill / outline colors for a given condition code}
\usage{
colors_for_condition(CONDITION_CODE = "AU")
}
\arguments{
\item{CONDITION_CODE}{A condition code}
}
\value{
a one-row data frame with three columns: \code{CONDITION_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for a given condition code
}
