% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_condition_types.R
\name{colors_for_condition_type}
\alias{colors_for_condition_type}
\title{Returns the fill / outline colors for a given condition type code}
\usage{
colors_for_condition_type(condition_type_code = "ALIVE")
}
\arguments{
\item{CONDITION_TYPE_CODE}{A condition type code}
}
\value{
a one-row data frame with three columns: \code{CONDITION_TYPE_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for a given condition type code
}
