% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_fates.R
\name{colors_for_fate}
\alias{colors_for_fate}
\title{Returns the fill / outline colors for a given fate code}
\usage{
colors_for_fate(fate_code = "RFL")
}
\arguments{
\item{fate_code}{A fate code}
}
\value{
a one-row data frame with three columns: \code{FATE_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for a given fate code
}
