% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_fate_types.R
\name{colors_for_fate_type}
\alias{colors_for_fate_type}
\title{Returns the fill / outline colors for a given fate type code}
\usage{
colors_for_fate_type(fate_type_code = "RETAINED")
}
\arguments{
\item{fate_type_code}{A fate type code}
}
\value{
a one-row data frame with three columns: \code{fate_type_code}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for a given fate type code
}
