% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_fisheries.R
\name{colors_for_fishery}
\alias{colors_for_fishery}
\title{Returns the fill / outline colors for a given fishery code (one among \code{PSLS}, \code{PSFS}, \code{LLD}, \code{LLF}, \code{BB}, \code{GN}, \code{LIH}, \code{LIT}, \code{LIC}, \code{OT})}
\usage{
colors_for_fishery(fishery_code = "PSLS")
}
\arguments{
\item{fishery_code}{A fishery code}
}
\value{
a one-row data frame with three columns: \code{FISHERY_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for a given fishery code (one among \code{PSLS}, \code{PSFS}, \code{LLD}, \code{LLF}, \code{BB}, \code{GN}, \code{LIH}, \code{LIT}, \code{LIC}, \code{OT})
}
