% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_fisheries.R
\name{colors_for_fishery_group}
\alias{colors_for_fishery_group}
\title{Returns the fill / outline colors for a given fishery group code (one among \code{PS}, \code{LL}, \code{BB}, \code{GN}, \code{LI}, \code{OT})}
\usage{
colors_for_fishery_group(fishery_group_code = "PS")
}
\arguments{
\item{fishery_group_code}{A fishery group code}
}
\value{
a one-row data frame with three columns: \code{FISHERY_GROUP_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for a given fishery group code (one among \code{PS}, \code{LL}, \code{BB}, \code{GN}, \code{LI}, \code{OT})
}
