% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_fishery_types.R
\name{colors_for_fishery_type}
\alias{colors_for_fishery_type}
\title{Returns the fill / outline colors for a given fishery type code (one among \code{ART} and \code{IND})}
\usage{
colors_for_fishery_type(fishery_type_code = "ART")
}
\arguments{
\item{fishery_type_code}{A fishery type code}
}
\value{
a one-row data frame with three columns: \code{FISHERY_TYPE_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for a given fishery type code (one among \code{ART} and \code{IND})
}
