% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_fishing_grounds.R
\name{colors_for_fishing_ground}
\alias{colors_for_fishing_ground}
\title{Returns the fill / outline colors for a given fishing ground code (one among \code{IRWESIO}, \code{IREASIO})}
\usage{
colors_for_fishing_ground(fishing_ground_code = "IRWESIO")
}
\arguments{
\item{fishing_ground_code}{A fishing ground code}
}
\value{
a one-row data frame with three columns: \code{FISHING_GROUND_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for a given fishing ground code (one among \code{IRWESIO}, \code{IREASIO})
}
