% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_fleets.R
\name{colors_for_fleet}
\alias{colors_for_fleet}
\title{Returns the fill / outline colors for a given fleet code}
\usage{
colors_for_fleet(fleet_code = "SYC")
}
\arguments{
\item{fleet_code}{A fleet code}
}
\value{
a one-row data frame with three columns: \code{FLEET_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for a given fleet code
}
