% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_gears.R
\name{colors_for_gear}
\alias{colors_for_gear}
\title{Returns the fill / outline colors for a given gear code (one among the 16 IOTC gear codes)}
\usage{
colors_for_gear(gear_code = "PS")
}
\arguments{
\item{gear_code}{A gear code}
}
\value{
a one-row data frame with three columns: \code{GEAR_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for a given gear code (one among the 16 IOTC gear codes)
}
