% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_IUCN.R
\name{colors_for_IUCN_status}
\alias{colors_for_IUCN_status}
\title{Returns the fill / outline colors for a given IUCN status code}
\usage{
colors_for_IUCN_status(IUCN_STATUS_CODE = "EN")
}
\arguments{
\item{IUCN_STATUS_CODE}{A IUCN status code}
}
\value{
a one-row data frame with three columns: \code{IUCN_STATUS_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for a given IUCN status code
}
