% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_quality_codes.R
\name{colors_for_quality}
\alias{colors_for_quality}
\title{Returns the fill / outline colors for a given quality code (one among \code{0}, \code{2}, \code{4}, \code{6} and \code{8})}
\usage{
colors_for_quality(quality_code = 0)
}
\arguments{
\item{quality_code}{A quality code}
}
\value{
a one-row data frame with three columns: \code{QUALITY_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for a given quality code (one among \code{0}, \code{2}, \code{4}, \code{6} and \code{8})
}
