% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_quarters.R
\name{colors_for_quarter}
\alias{colors_for_quarter}
\title{Returns the fill / outline colors for a given quarter}
\usage{
colors_for_quarter(quarter_code = "Q1")
}
\arguments{
\item{quarter_code}{A quarter}
}
\value{
a one-row data frame with three columns: \code{QUARTER}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for a given quarter
}
