% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_raisings.R
\name{colors_for_raising}
\alias{colors_for_raising}
\title{Returns the fill / outline colors for a given raise code (one among \code{OS}, \code{SD})}
\usage{
colors_for_raising(RAISE_CODE = "OS")
}
\arguments{
\item{RAISE_CODE}{A raise code}
}
\value{
a one-row data frame with three columns: \code{RAISE_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for a given raise code (one among \code{OS}, \code{SD})
}
