% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_species.R
\name{colors_for_species}
\alias{colors_for_species}
\title{Returns the fill / outline colors for a given species code (one among the 16 IOTC species codes)}
\usage{
colors_for_species(species_code = "YFT")
}
\arguments{
\item{species_code}{A species code}
}
\value{
a one-row data frame with three columns: \code{SPECIES_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for a given species code (one among the 16 IOTC species codes)
}
