% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_species_01_group.R
\name{colors_for_species_group}
\alias{colors_for_species_group}
\title{Returns the fill / outline colors for a given species group code (one among \code{SG_BILLFISH}, \code{SG_SEERFISH}, \code{SG_TUNAS}, \code{SG_SHARKS}, \code{SG_OTHER} and \code{OTHR})}
\usage{
colors_for_species_group(species_group_code = SG_TUNAS)
}
\arguments{
\item{species_group_code}{A species group code}
}
\value{
a one-row data frame with three columns: \code{SPECIES_GROUP_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for a given species group code (one among \code{SG_BILLFISH}, \code{SG_SEERFISH}, \code{SG_TUNAS}, \code{SG_SHARKS}, \code{SG_OTHER} and \code{OTHR})
}
