% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_species_00_wp.R
\name{colors_for_species_wp}
\alias{colors_for_species_wp}
\title{Returns the fill / outline colors for a given species WP code (one among \code{BILL}, \code{TEMP}, \code{TROP}, \code{NERI}, \code{BYCT} and \code{OTHR})}
\usage{
colors_for_species_wp(species_wp_code = WP_TROP)
}
\arguments{
\item{species_wp_code}{A species WP code}
}
\value{
a one-row data frame with three columns: \code{SPECIES_WP_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for a given species WP code (one among \code{BILL}, \code{TEMP}, \code{TROP}, \code{NERI}, \code{BYCT} and \code{OTHR})
}
