% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_fate_types.R
\name{fate_type_colors_for}
\alias{fate_type_colors_for}
\title{Returns the fill / outline colors for the fate types found within a data set}
\usage{
fate_type_colors_for(data)
}
\arguments{
\item{data}{A data set with a \code{fate_type_code} column}
}
\value{
a multiple rows data frame with three columns: \code{fate_type_code}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for the fate types found within a data set
}
