% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_conditions.R
\name{fill_for_condition}
\alias{fill_for_condition}
\title{Returns the fill color (as an RGB string) for a given condition code}
\usage{
fill_for_condition(CONDITION_CODE = "AU")
}
\arguments{
\item{CONDITION_CODE}{A condition code}
}
\value{
a string representing the RGB code of the fill color for the given condition
}
\description{
Returns the fill color (as an RGB string) for a given condition code
}
