% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_condition_types.R
\name{fill_for_condition_type}
\alias{fill_for_condition_type}
\title{Returns the fill color (as an RGB string) for a given condition type code}
\usage{
fill_for_condition_type(condition_type_code = "ALIVE")
}
\arguments{
\item{CONDITION_TYPE_CODE}{A condition type code}
}
\value{
a string representing the RGB code of the fill color for the given condition
}
\description{
Returns the fill color (as an RGB string) for a given condition type code
}
