% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_fate_types.R
\name{fill_for_fate_type}
\alias{fill_for_fate_type}
\title{Returns the fill color (as an RGB string) for a given fate type code}
\usage{
fill_for_fate_type(fate_type_code = "RETAINED")
}
\arguments{
\item{fate_type_code}{A fate type code}
}
\value{
a string representing the RGB code of the fill color for the given fate type
}
\description{
Returns the fill color (as an RGB string) for a given fate type code
}
