% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_fishing_grounds.R
\name{fill_for_fishing_ground}
\alias{fill_for_fishing_ground}
\title{Returns the fill color (as an RGB string) for a given fishing ground code (one among \code{IRWESIO}, \code{IREASIO})}
\usage{
fill_for_fishing_ground(fishing_ground_code = "IRWESIO")
}
\arguments{
\item{fishing_ground_code}{A fishing ground code}
}
\value{
a string representing the RGB code of the fill color for the given fishing ground
}
\description{
Returns the fill color (as an RGB string) for a given fishing ground code (one among \code{IRWESIO}, \code{IREASIO})
}
