% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_fleets.R
\name{fill_for_fleet}
\alias{fill_for_fleet}
\title{Returns the fill color (as an RGB string) for a given fleet code}
\usage{
fill_for_fleet(fleet_code = "SYC")
}
\arguments{
\item{fleet_code}{A fleet code}
}
\value{
a string representing the RGB code of the fill color for the given fleet
}
\description{
Returns the fill color (as an RGB string) for a given fleet code
}
