% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_gears.R
\name{fill_for_gear}
\alias{fill_for_gear}
\title{Returns the fill color (as an RGB string) for a given gear code}
\usage{
fill_for_gear(gear_code = "PS")
}
\arguments{
\item{gear_code}{A gear code}
}
\value{
a string representing the RGB code of the fill color for the given gear
}
\description{
Returns the fill color (as an RGB string) for a given gear code
}
