% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_IUCN.R
\name{fill_for_IUCN_status}
\alias{fill_for_IUCN_status}
\title{Returns the fill color (as an RGB string) for a given IUCN status code}
\usage{
fill_for_IUCN_status(IUCN_STATUS_CODE = "EN")
}
\arguments{
\item{IUCN_STATUS_CODE}{A IUCN status code}
}
\value{
a string representing the RGB code of the fill color for the given fate
}
\description{
Returns the fill color (as an RGB string) for a given IUCN status code
}
