% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_quality_codes.R
\name{fill_for_quality}
\alias{fill_for_quality}
\title{Returns the fill color (as an RGB string) for a given quality code (one among \code{0}, \code{2}, \code{4}, \code{6} and \code{8})}
\usage{
fill_for_quality(quality_code = 0)
}
\arguments{
\item{quality_code}{A quality code}
}
\value{
a string representing the RGB code of the fill color for the given quality code
}
\description{
Returns the fill color (as an RGB string) for a given quality code (one among \code{0}, \code{2}, \code{4}, \code{6} and \code{8})
}
