% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_quarters.R
\name{fill_for_quarter}
\alias{fill_for_quarter}
\title{Returns the fill color (as an RGB string) for a given quarter}
\usage{
fill_for_quarter(quarter_code = "Q1")
}
\arguments{
\item{quarter_code}{A quarter}
}
\value{
a string representing the RGB code of the fill color for the given quarter
}
\description{
Returns the fill color (as an RGB string) for a given quarter
}
