% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_raisings.R
\name{fill_for_raising}
\alias{fill_for_raising}
\title{Returns the fill color (as an RGB string) for a given raise code (one among \code{OS}, \code{SD})}
\usage{
fill_for_raising(RAISE_CODE = "OS")
}
\arguments{
\item{RAISE_CODE}{A raise code}
}
\value{
a string representing the RGB code of the fill color for the given raise code
}
\description{
Returns the fill color (as an RGB string) for a given raise code (one among \code{OS}, \code{SD})
}
