% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_species.R
\name{fill_for_species}
\alias{fill_for_species}
\title{Returns the fill color (as an RGB string) for a given species code (one among the 16 IOTC species codes)}
\usage{
fill_for_species(species_code = "YFT")
}
\arguments{
\item{species_code}{A species code}
}
\value{
a string representing the RGB code of the fill color for the given species
}
\description{
Returns the fill color (as an RGB string) for a given species code (one among the 16 IOTC species codes)
}
