% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_species_02_category.R
\name{fill_for_species_category}
\alias{fill_for_species_category}
\title{Returns the fill color (as an RGB string) for a given species category code (one among \code{BILLFISH}, \code{SEERFISH}, \code{NERITIC}, \code{TEMPERATE}, \code{TROPICAL}, \code{TUNAS_NEI}, \code{SHARKS}, \code{RAYS}, \code{SEABIRDS}, \code{TURTLES}, \code{CETACEANS}, \code{OTHERS})}
\usage{
fill_for_species_category(species_category_code = SC_TROPICAL)
}
\arguments{
\item{species_category_code}{A species category code}
}
\value{
a string representing the RGB code of the fill color for the given species category
}
\description{
Returns the fill color (as an RGB string) for a given species category code (one among \code{BILLFISH}, \code{SEERFISH}, \code{NERITIC}, \code{TEMPERATE}, \code{TROPICAL}, \code{TUNAS_NEI}, \code{SHARKS}, \code{RAYS}, \code{SEABIRDS}, \code{TURTLES}, \code{CETACEANS}, \code{OTHERS})
}
