% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_fishing_grounds.R
\name{fishing_ground_colors_for}
\alias{fishing_ground_colors_for}
\title{Returns the fill / outline colors for the fishing grounds found within a data set}
\usage{
fishing_ground_colors_for(data)
}
\arguments{
\item{data}{A data set with a \code{FISHING_GROUND_CODE} column}
}
\value{
a multiple rows data frame with three columns: \code{FISHING_GROUND_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for the fishing grounds found within a data set
}
