% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_fleets.R
\name{fleet_colors_for}
\alias{fleet_colors_for}
\title{Returns the fill / outline colors for the fleets found within a data set}
\usage{
fleet_colors_for(data)
}
\arguments{
\item{data}{A data set with a \code{FLEET_CODE} column}
}
\value{
a multiple rows data frame with three columns: \code{FLEET_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for the fleets found within a data set
}
