% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_helpers.R
\name{is_dark}
\alias{is_dark}
\title{Returns if the provided color is perceived as 'dark'}
\usage{
is_dark(color, threshold = 70)
}
\arguments{
\item{color}{A color}

\item{threshold}{a threshold to tell 'light' colors from 'dark' ones
(defaults to 70 but the link above suggests 123 instead, which is less effective, IMHO)}
}
\value{
\code{TRUE} if the color is perceived as 'dark', \code{FALSE} otherwise
}
\description{
See also: https://www.w3.org/TR/AERT#color-contrast
}
