% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_IUCN.R
\name{IUCN_colors_for}
\alias{IUCN_colors_for}
\title{Returns the fill / outline colors for the IUCN status codes found within a data set}
\usage{
IUCN_colors_for(data)
}
\arguments{
\item{data}{A data set with a \code{IUCN_STATUS_CODE} column}
}
\value{
a multiple rows data frame with three columns: \code{IUCN_STATUS_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for the IUCN status codes found within a data set
}
