% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_conditions.R
\name{outline_for_condition}
\alias{outline_for_condition}
\title{Returns the outline color (as an RGB string) for a given condition code}
\usage{
outline_for_condition(CONDITION_CODE = "AU")
}
\arguments{
\item{CONDITION_CODE}{A condition code}
}
\value{
a string representing the RGB code of the outline color for the given condition
}
\description{
Returns the outline color (as an RGB string) for a given condition code
}
