% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_fates.R
\name{outline_for_fate}
\alias{outline_for_fate}
\title{Returns the outline color (as an RGB string) for a given fate code}
\usage{
outline_for_fate(fate_code = "RFL")
}
\arguments{
\item{fate_code}{A fate code}
}
\value{
a string representing the RGB code of the outline color for the given fate
}
\description{
Returns the outline color (as an RGB string) for a given fate code
}
