% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_fate_types.R
\name{outline_for_fate_type}
\alias{outline_for_fate_type}
\title{Returns the outline color (as an RGB string) for a given fate type code}
\usage{
outline_for_fate_type(fate_type_code = "RETAINED")
}
\arguments{
\item{fate_type_code}{A fate type code}
}
\value{
a string representing the RGB code of the outline color for the given fate type
}
\description{
Returns the outline color (as an RGB string) for a given fate type code
}
