% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_fisheries.R
\name{outline_for_fishery}
\alias{outline_for_fishery}
\title{Returns the outline color (as an RGB string) for a given fishery code (one among \code{PSLS}, \code{PSFS}, \code{LLD}, \code{LLF}, \code{BB}, \code{GN}, \code{LIH}, \code{LIT}, \code{LIC}, \code{OT})}
\usage{
outline_for_fishery(fishery_code = "PSLS")
}
\arguments{
\item{fishery_code}{A fishery code}
}
\value{
a string representing the RGB code of the outline color for the given fishery
}
\description{
Returns the outline color (as an RGB string) for a given fishery code (one among \code{PSLS}, \code{PSFS}, \code{LLD}, \code{LLF}, \code{BB}, \code{GN}, \code{LIH}, \code{LIT}, \code{LIC}, \code{OT})
}
