% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_fisheries.R
\name{outline_for_fishery_group}
\alias{outline_for_fishery_group}
\title{Returns the outline color (as an RGB string) for a given fishery group code (one among \code{PS}, \code{LL}, \code{BB}, \code{GN}, \code{LI}, \code{OT})}
\usage{
outline_for_fishery_group(fishery_group_code = "PS")
}
\arguments{
\item{fishery_group_code}{A fishery group code}
}
\value{
a string representing the RGB code of the outline color for the given fishery group
}
\description{
Returns the outline color (as an RGB string) for a given fishery group code (one among \code{PS}, \code{LL}, \code{BB}, \code{GN}, \code{LI}, \code{OT})
}
