% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_fishery_types.R
\name{outline_for_fishery_type}
\alias{outline_for_fishery_type}
\title{Returns the outline color (as an RGB string) for a given fishery type code (one among \code{ART} and \code{IND})}
\usage{
outline_for_fishery_type(fishery_type_code = "ART")
}
\arguments{
\item{fishery_type_code}{A fishery type code}
}
\value{
a string representing the RGB code of the outline color for the given fishery type
}
\description{
Returns the outline color (as an RGB string) for a given fishery type code (one among \code{ART} and \code{IND})
}
