% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_fleets.R
\name{outline_for_fleet}
\alias{outline_for_fleet}
\title{Returns the outline color (as an RGB string) for a given fleet code}
\usage{
outline_for_fleet(fleet_code = "SYC")
}
\arguments{
\item{fishery_group_code}{A fleet code}
}
\value{
a string representing the RGB code of the outline color for the given fleet
}
\description{
Returns the outline color (as an RGB string) for a given fleet code
}
