% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_gears.R
\name{outline_for_gear}
\alias{outline_for_gear}
\title{Returns the outline color (as an RGB string) for a given gear code}
\usage{
outline_for_gear(gear_code = "PS")
}
\arguments{
\item{gear_code}{A gear code}
}
\value{
a string representing the RGB code of the outline color for the given gear
}
\description{
Returns the outline color (as an RGB string) for a given gear code
}
