% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_IUCN.R
\name{outline_for_IUCN_status}
\alias{outline_for_IUCN_status}
\title{Returns the outline color (as an RGB string) for a given IUCN status code}
\usage{
outline_for_IUCN_status(IUCN_STATUS_CODE = "EN")
}
\arguments{
\item{IUCN_STATUS_CODE}{A IUCN status code}
}
\value{
a string representing the RGB code of the outline color for the given fate
}
\description{
Returns the outline color (as an RGB string) for a given IUCN status code
}
