% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_raisings.R
\name{outline_for_raising}
\alias{outline_for_raising}
\title{Returns the outline color (as an RGB string) for a given raise code (one among \code{OS}, \code{SD})}
\usage{
outline_for_raising(RAISE_CODE = "OS")
}
\arguments{
\item{fishery_group_code}{A raise code}
}
\value{
a string representing the RGB code of the outline color for the given raise code
}
\description{
Returns the outline color (as an RGB string) for a given raise code (one among \code{OS}, \code{SD})
}
