% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_species_01_group.R
\name{outline_for_species_group}
\alias{outline_for_species_group}
\title{Returns the outline color (as an RGB string) for a given species group code (one among \code{SG_BILLFISH}, \code{SG_SEERFISH}, \code{SG_TUNAS}, \code{SG_SHARKS}, \code{SG_OTHER} and \code{OTHR})}
\usage{
outline_for_species_group(species_group_code = SG_TUNAS)
}
\arguments{
\item{species_group_code}{A species group code}
}
\value{
a string representing the RGB code of the outline color for the given species group
}
\description{
Returns the outline color (as an RGB string) for a given species group code (one among \code{SG_BILLFISH}, \code{SG_SEERFISH}, \code{SG_TUNAS}, \code{SG_SHARKS}, \code{SG_OTHER} and \code{OTHR})
}
