% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_species_00_wp.R
\name{outline_for_species_wp}
\alias{outline_for_species_wp}
\title{Returns the outline color (as an RGB string) for a given species WP code (one among \code{BILL}, \code{TEMP}, \code{TROP}, \code{NERI}, \code{BYCT} and \code{OTHR})}
\usage{
outline_for_species_wp(species_wp_code = WP_TROP)
}
\arguments{
\item{species_wp_code}{A species WP code}
}
\value{
a string representing the RGB code of the outline color for the given species WP
}
\description{
Returns the outline color (as an RGB string) for a given species WP code (one among \code{BILL}, \code{TEMP}, \code{TROP}, \code{NERI}, \code{BYCT} and \code{OTHR})
}
