% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_quality_codes.R
\name{quality_colors_for}
\alias{quality_colors_for}
\title{Returns the fill / outline colors for the quality codes found within a data set}
\usage{
quality_colors_for(data)
}
\arguments{
\item{data}{A data set with a \code{QUALITY_CODE} column}
}
\value{
a multiple rows data frame with three columns: \code{QUALITY_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for the quality codes found within a data set
}
