% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_quarters.R
\name{quarter_colors_for}
\alias{quarter_colors_for}
\title{Returns the fill / outline colors for the quarters found within a data set}
\usage{
quarter_colors_for(data)
}
\arguments{
\item{data}{A data set with a \code{QUARTER} column}
}
\value{
a multiple rows data frame with three columns: \code{QUARTER}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for the quarters found within a data set
}
