% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_raisings.R
\name{raising_colors_for}
\alias{raising_colors_for}
\title{Returns the fill / outline colors for the raise codes found within a data set}
\usage{
raising_colors_for(data)
}
\arguments{
\item{data}{A data set with a \code{RAISE_CODE} column}
}
\value{
a multiple rows data frame with three columns: \code{RAISE_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for the raise codes found within a data set
}
