% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_species.R
\name{reset_species_colors}
\alias{reset_species_colors}
\title{Resets the cached species colors table}
\usage{
reset_species_colors()
}
\description{
Resets the cached species colors table
}
