% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_species_01_group.R
\name{species_group_colors_for}
\alias{species_group_colors_for}
\title{Returns the fill / outline colors for the species groups found within a data set}
\usage{
species_group_colors_for(data)
}
\arguments{
\item{data}{A data set with a \code{SPECIES_GROUP_CODE} column}
}
\value{
a multiple rows data frame with three columns: \code{SPECIES_GROUP_CODE}, \code{FILL} and \code{OUTLINE}, with the latter two corresponding to the RGB values for the specific color
}
\description{
Returns the fill / outline colors for the species groups found within a data set
}
