% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_helpers.R
\name{unique_colors}
\alias{unique_colors}
\title{Returns a given number of unique colors from some of the most distinctive
palettes in RColorBrewer in the following order: \code{Paired}, \code{Dark2}, \code{Accent}, \code{Set1}, \code{Set2}, \code{Set3},
\code{Accent}, \code{Pastel1}, \code{Pastel2}}
\usage{
unique_colors(num_colors = 16)
}
\arguments{
\item{num_colors}{The number of colors to return}
}
\value{
an set of \code{num} colors extracted from the most distinctive palettes in RColorBrewer
}
\description{
Returns a given number of unique colors from some of the most distinctive
palettes in RColorBrewer in the following order: \code{Paired}, \code{Dark2}, \code{Accent}, \code{Set1}, \code{Set2}, \code{Set3},
\code{Accent}, \code{Pastel1}, \code{Pastel2}
}
