% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_gears.R
\name{update_cached_gear_colors}
\alias{update_cached_gear_colors}
\title{Return the cached gear colors table}
\usage{
update_cached_gear_colors(gear_colors)
}
\description{
Return the cached gear colors table
}
