% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_utils_aes_colors_species.R
\name{update_cached_species_colors}
\alias{update_cached_species_colors}
\title{Return the cached species colors table}
\usage{
update_cached_species_colors(species_colors)
}
\description{
Return the cached species colors table
}
